Release 1.1.0 - Add allValues method to retrieve values for a certain keyPath in a query. - Remove usage of throwing public functions. - Greatly reduce codebase size.

+++ Example/Tests/ManagedPerson.swift
@@ -25,7 +25,7 @@ extension ManagedPerson: RealmStorageManaging {
     
     public func storageObject() -> Person {
         
-        return Person(name: self.name, id: Int64(self.identifier))
+        return Person(name: self.name, id: self.identifier)
     }
     
     public static func fromStorageObject(_ item: Person) -> ManagedPerson {
